requirements.txt
click==8.0.3
colorama==0.4.4
dataclasses==0.8
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
PyJWT==2.3.0
SQLAlchemy==1.4.28
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
